move around einstein files
