+++ packages/template-retail-react-app/app/components/recommended-products/index.jsx
@@ -11,7 +11,7 @@ import {useIntl} from 'react-intl'
 import {Button} from '@chakra-ui/react'
 import ProductScroller from '../../components/product-scroller'
 import useCustomer from '../../commerce-api/hooks/useCustomer'
-import useEinstein from '../../hooks/useEinstein'
+import useEinstein from '../../hooks/use-einstein'
 import useIntersectionObserver from '../../hooks/use-intersection-observer'
 import useWishlist from '../../hooks/use-wishlist'
 import {useToast} from '../../hooks/use-toast'
