+++ packages/template-retail-react-app/app/hooks/einstein-mock-data.js
