+++ packages/template-retail-react-app/app/hooks/use-einstein.js
