+++ packages/template-retail-react-app/app/pages/home/index.jsx
@@ -35,7 +35,7 @@ import {getAssetUrl} from 'pwa-kit-react-sdk/ssr/universal/utils'
 import {heroFeatures, features} from './data'
 
 //Hooks
-import useEinstein from '../../hooks/useEinstein'
+import useEinstein from '../../hooks/use-einstein'
 
 // Constants
 import {
